terraform/functions/disable-billing/requirements.txt
functions-framework==3.*
google-cloud-billing==1.*
cloudevents==1.*
